Add trait injection through Traitable
